# api/openapi/objstore.yaml
openapi: 3.0.3
info:
  title: Go ObjectStore REST API
  description: |
    REST API for the go-objstore library providing unified access to multiple storage backends
    including local filesystem, AWS S3, Google Cloud Storage, and Azure Blob Storage.
  version: 0.1.0-beta
  contact:
    name: Go ObjectStore
  license:
    name: AGPL-3.0
    url: https://www.gnu.org/licenses/agpl-3.0.html
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: http://localhost:8080/api/v1
    description: Local development server (API v1)
...
paths:
...
  /objects:
    get:
      tags:
        - objects
      summary: List objects
      description: List objects with optional prefix filtering and pagination support
      operationId: listObjects
      parameters:
        - name: prefix
          in: query
          description: Filter objects by prefix
          required: false
          schema:
            type: string
            example: "documents/"
        - name: limit
          in: query
          description: Maximum number of results to return
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 100
            example: 50
        - name: token
          in: query
          description: Continuation token for pagination
          required: false
          schema:
            type: string
        - name: delimiter
          in: query
          description: Delimiter for hierarchical listing (e.g., "/" for directory-like structure)
          required: false
          schema:
            type: string
            example: "/"
      responses:
        '200':
          description: List of objects
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListObjectsResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /objects/{key}:
    put:
      tags:
        - objects
      summary: Upload object
      description: Upload an object to the storage backend with optional metadata
      operationId: putObject
      parameters:
        - name: key
          in: path
          description: Object key/path
          required: true
          schema:
            type: string
            example: "documents/file.pdf"
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: File to upload
                metadata:
                  type: string
                  description: JSON-encoded metadata object
                  example: '{"content_type":"application/pdf","custom":{"author":"John Doe"}}'
          application/octet-stream:
            schema:
              type: string
              format: binary
      responses:
        '201':
          description: Object uploaded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '413':
          description: Request entity too large
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    get:
      tags:
        - objects
      summary: Download object
      description: Retrieve an object from the storage backend
      operationId: getObject
      parameters:
        - name: key
          in: path
          description: Object key/path
          required: true
          schema:
            type: string
            example: "documents/file.pdf"
      responses:
        '200':
          description: Object content
          headers:
            Content-Type:
              schema:
                type: string
              description: MIME type of the object
            Content-Length:
              schema:
                type: integer
              description: Size of the object in bytes
            ETag:
              schema:
                type: string
              description: Entity tag for caching
            Last-Modified:
              schema:
                type: string
              description: Last modification timestamp
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '404':
          description: Object not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - objects
      summary: Delete object
      description: Remove an object from the storage backend
      operationId: deleteObject
      parameters:
        - name: key
          in: path
          description: Object key/path
          required: true
          schema:
            type: string
            example: "documents/file.pdf"
      responses:
        '204':
          description: Object deleted successfully
        '404':
          description: Object not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    head:
      tags:
        - objects
      summary: Get object metadata via HEAD
      description: >
        Check if an object exists and retrieve its metadata headers without
        downloading its content. Returns the same metadata headers as GET.
      operationId: headObject
      parameters:
        - name: key
          in: path
          description: Object key/path
          required: true
          schema:
            type: string
            example: "documents/file.pdf"
      responses:
        '200':
          description: Object exists
          headers:
            Content-Type:
              schema:
                type: string
              description: MIME type of the object
            Content-Length:
              schema:
                type: integer
              description: Size of the object in bytes
            ETag:
              schema:
                type: string
              description: Entity tag for caching
            Last-Modified:
              schema:
                type: string
              description: Last modification timestamp
        '404':
          description: Object not found
...
  /metadata/{key}:
    get:
      tags:
        - metadata
      summary: Get object metadata
      description: >
        Retrieve metadata for an object as a JSON body without downloading the
        object content. This route is parity with the QUIC GET /metadata/<key>
        route and the gRPC GetMetadata RPC.
      operationId: getObjectMetadata
      parameters:
        - name: key
          in: path
          description: Object key/path
          required: true
          schema:
            type: string
            example: "documents/file.pdf"
      responses:
        '200':
          description: Object metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ObjectResponse'
        '404':
          description: Object not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags:
        - metadata
      summary: Update object metadata
      description: Update metadata for an existing object
      operationId: updateObjectMetadata
      parameters:
        - name: key
          in: path
          description: Object key/path
          required: true
          schema:
            type: string
            example: "documents/file.pdf"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Metadata'
      responses:
        '200':
          description: Metadata updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Object not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /policies:
    get:
      tags:
        - lifecycle
      summary: List lifecycle policies
      description: Retrieve all configured lifecycle policies, optionally filtered by prefix
      operationId: getPolicies
      parameters:
        - name: prefix
          in: query
          description: Filter policies by object-key prefix
          required: false
          schema:
            type: string
      responses:
        '200':
          description: List of lifecycle policies
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PolicyListResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags:
        - lifecycle
      summary: Add lifecycle policy
      description: Create a new lifecycle policy
      operationId: addPolicy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PolicyRequest'
      responses:
        '201':
          description: Policy created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Policy already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /policies/{id}:
    delete:
      tags:
        - lifecycle
      summary: Remove lifecycle policy
      description: Delete a lifecycle policy by its ID
      operationId: removePolicy
      parameters:
        - name: id
          in: path
          description: Policy ID
          required: true
          schema:
            type: string
            example: "expire-old-docs"
      responses:
        '200':
          description: Policy removed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '404':
          description: Policy not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /policies/apply:
    post:
      tags:
        - lifecycle
      summary: Apply lifecycle policies
      description: >
        Evaluate all lifecycle policies against the current object set and
        execute the configured actions (delete or archive) for matching objects
        that have exceeded their retention period.
      operationId: applyPolicies
      responses:
        '200':
          description: Policies applied successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplyPoliciesResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
    ErrorResponse:
      type: object
      required:
        - error
        - code
      properties:
        error:
          type: string
          description: HTTP status text
          example: "Bad Request"
        code:
          type: integer
          description: HTTP status code
          example: 400
        message:
          type: string
          description: Detailed error message
          example: "key parameter is required"
    SuccessResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          description: Success message
          example: "operation completed successfully"
        data:
          type: object
          description: Optional response data
          additionalProperties: true
...
    ListObjectsResponse:
      type: object
      required:
        - objects
        - truncated
      properties:
        objects:
          type: array
          description: List of objects
          items:
            $ref: '#/components/schemas/ObjectResponse'
        common_prefixes:
          type: array
          description: Common prefixes (when using delimiter)
          items:
            type: string
          example:
            - "documents/"
            - "images/"
        next_token:
          type: string
          description: Token for fetching the next page
          example: "eyJrZXkiOiJkb2N1bWVudHMvZmlsZTEwMC5wZGYifQ=="
        truncated:
          type: boolean
          description: Whether more results are available
          example: false
...
    PolicyListResponse:
      type: object
      properties:
        policies:
          type: array
          items:
            $ref: '#/components/schemas/PolicyResponse'
        count:
          type: integer
          description: Total number of policies
          example: 3
    ApplyPoliciesResponse:
      type: object
      properties:
        message:
          type: string
          example: "lifecycle policies applied successfully"
        policies_count:
          type: integer
          description: Number of policies evaluated
          example: 2
        objects_processed:
          type: integer
          description: Number of objects affected
          example: 5
...